fix: Make the serialiser reusable.
